requirements.txt
python-telegram-bot
groq
google-generativeai
moviepy
